Changed debug terminal to use float variables

+++ include/config/robot_types.cpp
@@ -24,4 +24,4 @@ const RobotType robotIDTypes[] = {
 #endif
 
 //Goalie ID integer
-int GOALIE_ID = 5;
+float GOALIE_ID = 5;
